more updates

+++ app.py
@@ -6,8 +6,25 @@ import json
 import urllib.request #pip install urllib
 
 IMAGE_FORMATS = r'(?i)\.(jpg|png|gif|jpeg)$'
+
+def readCache():
+    try:
+        with open(cache_name, 'r') as filehandle:
+            for line in filehandle:
+                # remove linebreak which is the last character of the string
+                currentPlace = line[:-1]
+                # add item to the list
+                cache_set.add(currentPlace)
+    except FileNotFoundError:
+        print("No cache file found, creating new one")
+
+def writeCache():
+    with open(cache_name, 'w') as filehandle:
+        for listitem in cache_set:
+            filehandle.write('%s\n' % listitem)
+
 os.system('clear')
-print('Imagine Reddit Image Downloader\n' + "*" * 20 + "\n")
+print("*" * 35 + "\n* Imagine Reddit Image Downloader *\n" + "*" * 35 + "\n")
 
 # reads config file
 with open('config.json') as fi:
@@ -27,29 +44,14 @@ subreddit = praw_instance.subreddit(current_subreddit)
 
 cache_set = set()
 cache_name = config['settings']['cache_location'] +  "/" + current_subreddit + '.cache'
-
-try:
-    with open(cache_name, 'r') as filehandle:
-        for line in filehandle:
-            # remove linebreak which is the last character of the string
-            currentPlace = line[:-1]
-            # add item to the list
-            cache_set.add(currentPlace)
-except FileNotFoundError:
-    print("No cache file found, creating new one")
-
-def writeCache():
-    with open(cache_name, 'w') as filehandle:
-        for listitem in cache_set:
-            filehandle.write('%s\n' % listitem)
-
+readCache()
 
 for submission in subreddit.hot(limit=None):
     url = submission.url.rsplit('/',1)[-1]
-    if not(url in cache_set):
+    image_directory = config['settings']['cache_location'] + url # the directory for each image
+    if not(url in cache_set) and not(os.path.exists(image_directory)):
         if re.search(IMAGE_FORMATS, url): # starts download only for image files
             print(submission.title)
-            image_directory = config['settings']['cache_location'] + url # the directory for each image
             urllib.request.urlretrieve(submission.url, image_directory) # download image
             img = Image.open(image_directory) # create PIL image
             img.show()
@@ -65,6 +67,7 @@ for submission in subreddit.hot(limit=None):
                 save_fi = input("Save current image (y/n): ")
                 if (save_fi == 'n'):
                     os.system('rm ' + image_directory)
+                    cache_set.add(url)
                 writeCache()
                 exit()
